+++ lib/flutter_map.dart
@@ -59,5 +59,3 @@ export 'package:flutter_map/src/map/options/cursor_keyboard_rotation.dart';
 export 'package:flutter_map/src/map/options/interaction.dart';
 export 'package:flutter_map/src/map/options/options.dart';
 export 'package:flutter_map/src/map/widget.dart';
-export 'package:flutter_map/src/misc/bounds.dart';
-export 'package:flutter_map/src/misc/extensions.dart';