+++ lib/src/layer/marker_layer/marker_layer.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:flutter/widgets.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
@@ -49,21 +47,24 @@ class MarkerLayer extends StatelessWidget {
         children: (List<Marker> markers) sync* {
           for (final m in markers) {
             // Resolve real alignment
+            // TODO this can probably just be done with calls to Size, Offset, and Rect
             final left = 0.5 * m.width * ((m.alignment ?? alignment).x + 1);
             final top = 0.5 * m.height * ((m.alignment ?? alignment).y + 1);
             final right = m.width - left;
             final bottom = m.height - top;
 
             // Perform projection
-            final pxPoint = map.project(m.point);
+            final pxPoint = map.projectAtZoom(m.point);
 
             // Cull if out of bounds
-            if (!map.pixelBounds.containsPartialBounds(
-              Bounds(
-                Point(pxPoint.x + left, pxPoint.y - bottom),
-                Point(pxPoint.x - right, pxPoint.y + top),
+            if (!map.pixelBounds.overlaps(
+              Rect.fromPoints(
+                Offset(pxPoint.dx + left, pxPoint.dy - bottom),
+                Offset(pxPoint.dx - right, pxPoint.dy + top),
               ),
-            )) continue;
+            )) {
+              continue;
+            }
 
             // Apply map camera to marker position
             final pos = pxPoint - map.pixelOrigin;
@@ -72,8 +73,8 @@ class MarkerLayer extends StatelessWidget {
               key: m.key,
               width: m.width,
               height: m.height,
-              left: pos.x - right,
-              top: pos.y - bottom,
+              left: pos.dx - right,
+              top: pos.dy - bottom,
               child: (m.rotate ?? rotate)
                   ? Transform.rotate(
                       angle: -map.rotationRad,