+++ lib/src/layer/polyline_layer/painter.dart
@@ -20,7 +20,7 @@ base class _PolylinePainter<R extends Object>
   @override
   bool elementHitTest(
     _ProjectedPolyline<R> projectedPolyline, {
-    required math.Point<double> point,
+    required Offset point,
     required LatLng coordinate,
   }) {
     final polyline = projectedPolyline.polyline;
@@ -57,7 +57,7 @@ base class _PolylinePainter<R extends Object>
       final o2 = offsets[i + 1];
 
       final distanceSq =
-          getSqSegDist(point.x, point.y, o1.dx, o1.dy, o2.dx, o2.dy);
+          getSqSegDist(point.dx, point.dy, o1.dx, o1.dy, o2.dx, o2.dy);
 
       if (distanceSq <= hittableDistance * hittableDistance) return true;
     }
@@ -108,7 +108,7 @@ base class _PolylinePainter<R extends Object>
     }
 
     final origin =
-        camera.project(camera.center).toOffset() - camera.size.toOffset() / 2;
+        camera.projectAtZoom(camera.center) - camera.size.center(Offset.zero);
 
     for (final projectedPolyline in polylines) {
       final polyline = projectedPolyline.polyline;
@@ -284,8 +284,8 @@ base class _PolylinePainter<R extends Object>
     return delta.distance;
   }
 
-  LatLng _unproject(DoublePoint p0) =>
-      camera.crs.projection.unprojectXY(p0.x, p0.y);
+  LatLng _unproject(Offset p0) =>
+      camera.crs.projection.unprojectXY(p0.dx, p0.dy);
 
   @override
   bool shouldRepaint(_PolylinePainter<R> oldDelegate) =>