+++ lib/src/layer/polyline_layer/projected_polyline.dart
@@ -3,7 +3,7 @@ part of 'polyline_layer.dart';
 @immutable
 class _ProjectedPolyline<R extends Object> with HitDetectableElement<R> {
   final Polyline<R> polyline;
-  final List<DoublePoint> points;
+  final List<Offset> points;
 
   @override
   R? get hitValue => polyline.hitValue;