+++ test/map/camera/camera_constraint_test.dart
@@ -1,4 +1,4 @@
-import 'dart:math';
+import 'dart:ui';
 
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -20,7 +20,7 @@ void main() {
           center: const LatLng(-90, -180),
           zoom: 1,
           rotation: 45,
-          nonRotatedSize: const Point(200, 300),
+          nonRotatedSize: const Size(200, 300),
         );
 
         final clamped = mapConstraint.constrain(camera)!;